refactor: renaming and structure
