+++ .eslintrc.json
@@ -14,6 +14,7 @@
     "no-bitwise": 0,
     "prettier/prettier": 2,
     "no-unused-expressions": 0,
+    "react/forbid-prop-types": 0,
     "import/no-extraneous-dependencies": [2, { "devDependencies": true }],
     "prefer-destructuring": ["error", { "object": true, "array": false }]
   },
