+++ package.json
@@ -53,6 +53,7 @@
     "eslint-plugin-prettier": "3.1.2",
     "eslint-plugin-react": "7.13.0",
     "prettier": "1.19.1",
+    "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-proptypes": "^1.0.0",
