+++ src/table/cells/icon.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 export default function withIcon(CellComponent) {
   class HOC extends React.PureComponent {
@@ -7,5 +8,10 @@ export default function withIcon(CellComponent) {
       return <CellComponent {...this.props}>{`⚜ ${value}`}</CellComponent>;
     }
   }
+
+  HOC.propTypes = {
+    value: PropTypes.string.isRequired,
+  };
+
   return HOC;
 }
